+ moved refactored alphazero.node.* into alphazero + moved preprocessor.json into molecule_game
